+++ packages/bfx-containers/src/redux/actions/index.js
@@ -1,1 +1,2 @@
 export * from './ws.actions'
+export * from './ticker.actions'