+++ packages/bfx-containers/src/redux/sagas/index.js
@@ -3,9 +3,10 @@ import { all } from 'redux-saga/effects'
 import bootstrapSaga from './bootstrap.saga'
 import currenciesSaga from './currencies.saga'
 import notificationsSaga from './notifications.saga'
+import tickersSaga from './ticker.saga'
 import wsSaga from './ws.saga'
 
-export { wsSaga }
+export { wsSaga, tickersSaga }
 
 export default function* rootSaga() {
   // add common non-injected sagas here...