+++ packages/bfx-containers/src/redux/selectors/index.js
@@ -1,3 +1,5 @@
 export * from './ws.selectors'
 export * from './book.selectors'
 export * from './trades.selectors'
+export * from './ticker.selectors'
+export * from './currencies.selectors'