+++ packages/core/src/components/DepthChart/DepthChart.helpers.js
@@ -30,12 +30,12 @@ try {
 } catch (e) {}
 
 const wheelOpt = supportsPassive ? { passive: false } : false
-const wheelEvent = 'onwheel' in document.createElement('div') ? 'wheel' : 'mousewheel'
+const getWheelEvent = () => ('onwheel' in document.createElement('div') ? 'wheel' : 'mousewheel')
 
 // call this to Disable
 export function disableScroll() {
   window.addEventListener('DOMMouseScroll', preventDefault, false) // older FF
-  window.addEventListener(wheelEvent, preventDefault, wheelOpt) // modern desktop
+  window.addEventListener(getWheelEvent(), preventDefault, wheelOpt) // modern desktop
   window.addEventListener('touchmove', preventDefault, wheelOpt) // mobile
   window.addEventListener('keydown', preventDefaultForScrollKeys, false)
 }
@@ -43,7 +43,7 @@ export function disableScroll() {
 // call this to Enable
 export function enableScroll() {
   window.removeEventListener('DOMMouseScroll', preventDefault, false)
-  window.removeEventListener(wheelEvent, preventDefault, wheelOpt)
+  window.removeEventListener(getWheelEvent(), preventDefault, wheelOpt)
   window.removeEventListener('touchmove', preventDefault, wheelOpt)
   window.removeEventListener('keydown', preventDefaultForScrollKeys, false)
 }