+++ packages/core/src/components/TickerList/TickerList.Header/index.js
@@ -1,1 +1,2 @@
-export { default } from './TickerList.Header'
+export { default as Favorite } from './Favorite'
+export { default as Volume } from './Volume'