+++ packages/core/src/components/TickerList/TickerList.Toolbar.js
@@ -1,27 +1,33 @@
-import { faSearch } from '@fortawesome/free-solid-svg-icons'
+import { faSearch, faTimes } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import PropTypes from 'prop-types'
-import React, { memo } from 'react'
+import React, { memo, useCallback } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { Input } from '../ui'
+import { Input, Button } from '../ui'
 
 const TickerListToolbar = (props) => {
-  const {
-    searchTerm,
-    setSearchTerm,
-  } = props
+  const { searchTerm, setSearchTerm } = props
   const { t } = useTranslation()
+  const onCancelClick = useCallback(() => setSearchTerm(''), [setSearchTerm])
 
   return (
     <Input
       placeholder={`${t('common:search')}..`}
       small
-      rightElement={<FontAwesomeIcon icon={faSearch} className='search-icon' />}
+      rightElement={
+        searchTerm ? (
+          <Button onClick={onCancelClick} minimal>
+            <FontAwesomeIcon icon={faTimes} className='search-icon' />
+          </Button>
+        ) : (
+          <FontAwesomeIcon icon={faSearch} className='search-icon' />
+        )
+      }
       value={searchTerm}
       className='search-ccy'
       onChange={setSearchTerm}
-      type='search'
+      type='text'
     />
   )
 }