+++ packages/core/src/components/TickerList/stories/TickerList.stories_data.js
@@ -6,6 +6,7 @@ const tickerList = [
     changePerc: -0.006,
     lastPrice: 4.019,
     volume: 402,
+    ccyLabels: ['UST', 'Tether USDt', 'EOS'],
   }, {
     id: 'tBTCUSD',
     baseCcy: 'BTC',
@@ -13,6 +14,7 @@ const tickerList = [
     changePerc: -0.0006,
     lastPrice: 9000,
     volume: 3940,
+    ccyLabels: ['BTC', 'Bitcoin', 'USD', 'US Dollar'],
   }, {
     id: 'tBTCUST',
     baseCcy: 'BTC',
@@ -20,6 +22,16 @@ const tickerList = [
     changePerc: 0,
     lastPrice: 200,
     volume: 10000,
+    ccyLabels: ['BTC', 'Bitcoin', 'UST', 'Tether USDt'],
+  },
+  {
+    id: 'tXLMF0:USTF0',
+    baseCcy: 'tXLMF0',
+    quoteCcy: 'USTF0',
+    changePerc: 2,
+    lastPrice: 20,
+    volume: 5587,
+    ccyLabels: ['tXLMF0', 'USTF0'],
   },
 ]