+++ packages/storybook/.storybook/main.js
@@ -8,6 +8,11 @@ module.exports = {
     '../../bfx-containers/src/containers/OrderHistory/**/*.stories.js',
     '../../core/src/components/Trades/**/*.stories.js',
     '../../bfx-containers/src/containers/Trades/**/*.stories.js',
+    '../../core/src/components/Ticker/**/*.stories.js',
+    '../../bfx-containers/src/containers/Ticker/**/*.stories.js',
+    '../../core/src/components/TickerList/**/*.stories.js',
+    '../../bfx-containers/src/containers/TickerList/**/*.stories.js',
+    '../../core/src/components/ui/VirtualTable/**/*.stories.js',
   ],
   addons: [
     '@storybook/preset-scss',